added EquipmentTableFrame
